+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+1.1.0 - November 26, 2014
+
+* Add option to select what video IDs in a videoset to download
+* Tweaks to argument names
+
+1.0.1 - November 8, 2014
+
+* Slight improvements to help text and README
+
 1.0.0 - November 7, 2014
 
 * First public release, compatible with CBox 3.0.2.4